+++ src/detail/zstd_stream.cpp
@@ -4,11 +4,10 @@
 #include <sstream>
 #include <utility>  // move
 
-#include "databento/detail/buffer.hpp"
 #include "databento/exceptions.hpp"
-#include "databento/log.hpp"
 
 using databento::detail::ZstdDecodeStream;
+using Status = databento::IReadable::Status;
 
 ZstdDecodeStream::ZstdDecodeStream(std::unique_ptr<IReadable> input)
     : input_{std::move(input)},
@@ -42,8 +41,14 @@ void ZstdDecodeStream::ReadExact(std::byte* buffer, std::size_t length) {
 }
 
 std::size_t ZstdDecodeStream::ReadSome(std::byte* buffer, std::size_t max_length) {
+  return ReadSome(buffer, max_length, std::chrono::milliseconds{}).read_size;
+}
+
+databento::IReadable::Result ZstdDecodeStream::ReadSome(
+    std::byte* buffer, std::size_t max_length, std::chrono::milliseconds timeout) {
   ZSTD_outBuffer z_out_buffer{buffer, max_length, 0};
-  std::size_t read_size = 0;
+  databento::IReadable::Result read_result{0, Status::Ok};
+
   do {
     const auto unread_input = z_in_buffer_.size - z_in_buffer_.pos;
     if (unread_input > 0) {
@@ -59,9 +64,18 @@ std::size_t ZstdDecodeStream::ReadSome(std::byte* buffer, std::size_t max_length
       in_buffer_.resize(new_size);
       z_in_buffer_.src = in_buffer_.data();
     }
-    read_size = input_->ReadSome(&in_buffer_[unread_input], read_suggestion_);
-    z_in_buffer_.size = unread_input + read_size;
+
+    // Only apply timeout to inner reader for simplicity
+    read_result =
+        input_->ReadSome(&in_buffer_[unread_input], read_suggestion_, timeout);
+    z_in_buffer_.size = unread_input + read_result.read_size;
     z_in_buffer_.pos = 0;
+
+    // No data to decompress: timeout or closed. Calling `ZSTD_decompressStream` with an
+    // empty input and no buffered data will trigger an error
+    if (z_in_buffer_.size == 0) {
+      break;
+    }
 
     read_suggestion_ =
         ::ZSTD_decompressStream(z_dstream_.get(), &z_out_buffer, &z_in_buffer_);
@@ -69,8 +83,11 @@ std::size_t ZstdDecodeStream::ReadSome(std::byte* buffer, std::size_t max_length
       throw DbnResponseError{std::string{"Zstd error decompressing: "} +
                              ::ZSTD_getErrorName(read_suggestion_)};
     }
-  } while (z_out_buffer.pos == 0 && read_size > 0);
-  return z_out_buffer.pos;
+  } while (z_out_buffer.pos == 0 && read_result.read_size > 0);
+
+  const auto read_size = z_out_buffer.pos;
+  // Only return inner read status if there's no data
+  return {read_size, read_size > 0 ? Status::Ok : read_result.status};
 }
 
 using databento::detail::ZstdCompressStream;
@@ -91,27 +108,7 @@ ZstdCompressStream::ZstdCompressStream(ILogReceiver* log_receiver, IWritable* ou
   ::ZSTD_CCtx_setParameter(z_cstream_.get(), ZSTD_c_checksumFlag, 1);
 }
 
-ZstdCompressStream::~ZstdCompressStream() {
-  ZSTD_outBuffer z_out_buffer{out_buffer_.data(), out_buffer_.size(), 0};
-  while (true) {
-    const std::size_t remaining = ::ZSTD_compressStream2(
-        z_cstream_.get(), &z_out_buffer, &z_in_buffer_, ::ZSTD_e_end);
-    if (remaining == 0) {
-      break;
-    }
-    if (::ZSTD_isError(remaining) && log_receiver_) {
-      log_receiver_->Receive(LogLevel::Error,
-                             std::string{"Zstd error compressing end of stream: "} +
-                                 ::ZSTD_getErrorName(remaining));
-      break;
-    }
-  }
-  assert(z_in_buffer_.pos == z_in_buffer_.size);
-  // Forward compressed output
-  if (z_out_buffer.pos > 0) {
-    output_->WriteAll(out_buffer_.data(), z_out_buffer.pos);
-  }
-}
+ZstdCompressStream::~ZstdCompressStream() { Flush(); }
 
 void ZstdCompressStream::WriteAll(const std::byte* buffer, std::size_t length) {
   in_buffer_.insert(in_buffer_.end(), buffer, buffer + length);
@@ -137,4 +134,29 @@ void ZstdCompressStream::WriteAll(const std::byte* buffer, std::size_t length) {
       output_->WriteAll(out_buffer_.data(), z_out_buffer.pos);
     }
   }
+}
+
+void ZstdCompressStream::Flush() {
+  ZSTD_outBuffer z_out_buffer{out_buffer_.data(), out_buffer_.size(), 0};
+  while (true) {
+    const std::size_t remaining = ::ZSTD_compressStream2(
+        z_cstream_.get(), &z_out_buffer, &z_in_buffer_, ::ZSTD_e_end);
+    if (remaining == 0) {
+      break;
+    }
+    if (::ZSTD_isError(remaining) && log_receiver_) {
+      log_receiver_->Receive(LogLevel::Error,
+                             std::string{"Zstd error compressing end of stream: "} +
+                                 ::ZSTD_getErrorName(remaining));
+      break;
+    }
+  }
+  assert(z_in_buffer_.pos == z_in_buffer_.size);
+  // Forward compressed output
+  if (z_out_buffer.pos > 0) {
+    output_->WriteAll(out_buffer_.data(), z_out_buffer.pos);
+  }
+  // Clear the input buffer since it's all been flushed
+  in_buffer_.clear();
+  z_in_buffer_ = {in_buffer_.data(), 0, 0};
 }